repaired: now compiling! delete unnecessary userInput states TODO: make unittests compile
